Revert "Configuration to avoid fetching resource (#203)"
This reverts commit f46ea65780afeff110a7dadd991c10e34552ef0d.
