+++ microcosm_pubsub/dispatcher.py
@@ -44,8 +44,7 @@ from microcosm_pubsub.result import MessageHandlingResult, MessageHandlingResult
 @logger
 @defaults(
     # Number of failed attempts after which the message stops being processed
-    message_max_processing_attempts=typed(int, default_value=None),
-    fetch_uri_resource=typed(bool, default_value=True),
+    message_max_processing_attempts=typed(int, default_value=None)
 )
 class SQSMessageDispatcher:
     """
