+++ setup.py
@@ -25,7 +25,6 @@ setup(
         "microcosm-caching>=0.2.0",
         "microcosm-daemon>=1.0.0",
         "microcosm-logging>=1.3.0",
-        "tornado<6",
     ],
     extras_require={
         "metrics": "microcosm-metrics>=2.5.0",
